+++ src/libraries/TypesLib.sol
@@ -23,7 +23,7 @@ library TypesLib {
         uint256 subId; // must be 0 for direct funding
         uint256 directFundingFeePaid; // must be > 0 for direct funding and if subId == 0
         uint32 callbackGasLimit; // must be between 0 and maxGasLimit
-        uint64 requestId;
+        uint256 requestId;
         bytes message;
         bytes condition;
         bytes signature;
