+++ src/mocks/MockRandomnessReceiver.sol
@@ -11,7 +11,7 @@ contract MockRandomnessReceiver is RandomnessReceiverBase {
     bytes32 public randomness;
 
     /// @notice Stores the request ID of the latest randomness request
-    uint64 public requestId;
+    uint256 public requestId;
 
     /// @notice Initializes the contract with the address of the randomness sender
     /// @param randomnessSender The address of the randomness provider
@@ -19,9 +19,9 @@ contract MockRandomnessReceiver is RandomnessReceiverBase {
 
     /// @notice Requests randomness using the direct funding option
     /// @dev Calls `_requestRandomnessPayInNative` to get a random value, updating `requestId` with the request ID
-    function rollDiceWithDirectFunding(uint32 callbackGasLimit) external returns (uint64, uint256) {
+    function rollDiceWithDirectFunding(uint32 callbackGasLimit) external returns (uint256, uint256) {
         // create randomness request
-        (uint64 requestID, uint256 requestPrice) = _requestRandomnessPayInNative(callbackGasLimit);
+        (uint256 requestID, uint256 requestPrice) = _requestRandomnessPayInNative(callbackGasLimit);
         // store request id
         requestId = requestID;
         return (requestID, requestPrice);
@@ -29,9 +29,9 @@ contract MockRandomnessReceiver is RandomnessReceiverBase {
 
     /// @notice Requests randomness using the subscription option
     /// @dev Calls `_requestRandomnessWithSubscription` to get a random value, updating `requestId` with the request ID
-    function rollDiceWithSubscription(uint32 callbackGasLimit) external payable returns (uint64) {
+    function rollDiceWithSubscription(uint32 callbackGasLimit) external payable returns (uint256) {
         // create randomness request
-        uint64 requestID = _requestRandomnessWithSubscription(callbackGasLimit);
+        uint256 requestID = _requestRandomnessWithSubscription(callbackGasLimit);
         // store request id
         requestId = requestID;
         return requestID;
@@ -45,7 +45,7 @@ contract MockRandomnessReceiver is RandomnessReceiverBase {
     /// @dev Ensures the received request ID matches the stored one before updating state
     /// @param requestID The ID of the randomness request
     /// @param _randomness The random value received from the oracle
-    function onRandomnessReceived(uint64 requestID, bytes32 _randomness) internal override {
+    function onRandomnessReceived(uint256 requestID, bytes32 _randomness) internal override {
         require(requestId == requestID, "Request ID mismatch");
         randomness = _randomness;
     }
